% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_selector.R
\name{all_selector}
\alias{all_selector}
\title{All Voxels Series Selector}
\usage{
all_selector()
}
\value{
An object of class \code{all_selector}
}
\description{
Select all voxels in the dataset mask.
}
\examples{
# Select all voxels
sel <- all_selector()
}
