% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_series_resolvers.R
\name{all_timepoints}
\alias{all_timepoints}
\title{Helper returning all timepoints for a dataset}
\usage{
all_timepoints(dataset)
}
\arguments{
\item{dataset}{An \code{fmri_dataset} object.}
}
\value{
Integer vector of all valid timepoint indices.
}
\description{
Helper returning all timepoints for a dataset
}
\keyword{internal}
