% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_helpers.R
\name{analyze_run}
\alias{analyze_run}
\title{Simulate Analysis Function for Examples}
\usage{
analyze_run(data, method = "mean")
}
\arguments{
\item{data}{Data matrix}

\item{method}{Analysis method}
}
\value{
Numeric vector of results
}
\description{
Simulate Analysis Function for Examples
}
\keyword{internal}
