% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/dataset_methods.R,
%   R/sampling_frame_adapters.R
\name{blockids}
\alias{blockids}
\alias{blockids.matrix_dataset}
\alias{blockids.fmri_dataset}
\alias{blockids.fmri_mem_dataset}
\alias{blockids.fmri_file_dataset}
\alias{blockids.fmri_study_dataset}
\alias{blockids.sampling_frame}
\title{Get Block IDs from Sampling Frame}
\usage{
blockids(x, ...)

\method{blockids}{matrix_dataset}(x, ...)

\method{blockids}{fmri_dataset}(x, ...)

\method{blockids}{fmri_mem_dataset}(x, ...)

\method{blockids}{fmri_file_dataset}(x, ...)

\method{blockids}{fmri_study_dataset}(x, ...)

\method{blockids}{sampling_frame}(x, ...)
}
\arguments{
\item{x}{An object containing temporal structure (e.g., sampling_frame, fmri_dataset)}

\item{...}{Additional arguments passed to methods}
}
\value{
Integer vector of length equal to total timepoints, with
values indicating run membership (1 for first run, 2 for second, etc.)
}
\description{
Generates a vector of block/run identifiers for each timepoint.
}
\details{
This function creates a vector where each element indicates which
run/block the corresponding timepoint belongs to. This is useful
for run-wise analyses or modeling run effects.
}
\examples{
\donttest{
# Create a sampling frame with 2 runs of different lengths
sf <- fmrihrf::sampling_frame(blocklens = c(3, 4), TR = 2)
blockids(sf)  # Returns: c(1, 1, 1, 2, 2, 2, 2)
}

}
\seealso{
\code{\link{get_run_lengths}} for run lengths,
\code{\link{samples}} for timepoint indices
}
