% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{blocklens}
\alias{blocklens}
\title{Get Block Lengths from Objects}
\usage{
blocklens(x, ...)
}
\arguments{
\item{x}{An object with block structure}

\item{...}{Additional arguments passed to methods}
}
\value{
Integer vector of block/run lengths
}
\description{
Generic function to extract block/run lengths from various objects.
Extends the sampling_frame generic to work with dataset objects.
}
