% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_backend_seed_s3.R
\name{chunkGrid.study_backend_seed}
\alias{chunkGrid.study_backend_seed}
\title{Get Chunk Grid for Study Backend Seed}
\usage{
chunkGrid.study_backend_seed(x, chunk_dim = NULL)
}
\arguments{
\item{x}{A study_backend_seed object}

\item{chunk_dim}{Optional chunk dimensions}
}
\value{
A RegularArrayGrid object
}
\description{
Get Chunk Grid for Study Backend Seed
}
