% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chunks.R
\name{data_chunks.fmri_study_dataset}
\alias{data_chunks.fmri_study_dataset}
\title{Create Data Chunks for fmri_study_dataset Objects}
\usage{
\method{data_chunks}{fmri_study_dataset}(x, nchunks = 1, runwise = FALSE, ...)
}
\arguments{
\item{x}{An object of class 'fmri_study_dataset'}

\item{nchunks}{The number of chunks to split the data into. Default is 1.}

\item{runwise}{If TRUE, creates run-wise chunks instead of arbitrary chunks}

\item{...}{Additional arguments passed to methods}
}
\value{
A list of data chunks, each containing data, indices and chunk number
}
\description{
This function creates data chunks for multi-subject study datasets.
}
