% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_delayed_array.R
\name{dim,StorageBackendSeed-method}
\alias{dim,StorageBackendSeed-method}
\title{Dimensions of StorageBackendSeed}
\usage{
\S4method{dim}{StorageBackendSeed}(x)
}
\arguments{
\item{x}{A StorageBackendSeed object}
}
\value{
An integer vector of length 2 (timepoints, voxels)
}
\description{
Dimensions of StorageBackendSeed
}
\keyword{internal}
