% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_backend_seed.R
\name{dim,StudyBackendSeed-method}
\alias{dim,StudyBackendSeed-method}
\title{Dimensions of StudyBackendSeed}
\usage{
\S4method{dim}{StudyBackendSeed}(x)
}
\arguments{
\item{x}{A StudyBackendSeed object}
}
\value{
Integer vector of dimensions
}
\description{
Dimensions of StudyBackendSeed
}
\keyword{internal}
