% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FmriSeries.R
\name{dim.fmri_series}
\alias{dim.fmri_series}
\title{Dimensions of fmri_series}
\usage{
\method{dim}{fmri_series}(x)
}
\arguments{
\item{x}{An fmri_series object}
}
\value{
Integer vector of length 2 (timepoints, voxels)
}
\description{
Dimensions of fmri_series
}
