% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_backend_seed.R
\name{dimnames,StudyBackendSeed-method}
\alias{dimnames,StudyBackendSeed-method}
\title{Dimnames of StudyBackendSeed}
\usage{
\S4method{dimnames}{StudyBackendSeed}(x)
}
\arguments{
\item{x}{A StudyBackendSeed object}
}
\value{
List of dimnames (always NULL for this class)
}
\description{
Dimnames of StudyBackendSeed
}
\keyword{internal}
