% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_backend.R
\name{.get_offset}
\alias{.get_offset}
\title{Get offset from a LatentNeuroVec object}
\usage{
.get_offset(obj)
}
\arguments{
\item{obj}{A LatentNeuroVec object}
}
\value{
Numeric vector (may be length 0)
}
\description{
Get offset from a LatentNeuroVec object
}
\keyword{internal}
