% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_h5_backend.R
\name{.has_shared_template}
\alias{.has_shared_template}
\title{Check if archive has a shared template}
\usage{
.has_shared_template(h5_handle)
}
\arguments{
\item{h5_handle}{An open \code{hdf5r::H5File} object.}
}
\value{
Logical.
}
\description{
Check if archive has a shared template
}
\keyword{internal}
