% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_h5_write.R
\name{.make_scan_name}
\alias{.make_scan_name}
\title{Build scan name string from scan manifest row}
\usage{
.make_scan_name(scans_df)
}
\description{
Build scan name string from scan manifest row
}
\keyword{internal}
