% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_h5_write.R
\name{.read_scan_confounds}
\alias{.read_scan_confounds}
\title{Read confounds for a single scan from a BIDS project}
\usage{
.read_scan_confounds(bids_proj, scan_row, confounds_spec)
}
\description{
Read confounds for a single scan from a BIDS project
}
\keyword{internal}
