% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_h5_write.R
\name{.read_scan_events}
\alias{.read_scan_events}
\title{Read events for a single scan from a BIDS project}
\usage{
.read_scan_events(bids_proj, scan_row)
}
\description{
Read events for a single scan from a BIDS project
}
\keyword{internal}
