% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_h5_backend.R
\name{.read_template_loadings}
\alias{.read_template_loadings}
\title{Read shared template loadings}
\usage{
.read_template_loadings(h5_handle)
}
\arguments{
\item{h5_handle}{An open \code{hdf5r::H5File} object.}
}
\value{
A numeric matrix \code{[V, K]} or \code{NULL}.
}
\description{
Read shared template loadings
}
\keyword{internal}
