% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.require_namespace}
\alias{.require_namespace}
\title{Internal wrapper for requireNamespace}
\usage{
.require_namespace(package, ...)
}
\description{
Centralises optional dependency checks so tests can mock behaviour.
}
\keyword{internal}
