% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_h5_write.R
\name{.write_bids_group}
\alias{.write_bids_group}
\title{Write /bids/ group to HDF5 file}
\usage{
.write_bids_group(h5, bids_proj, scans_df, space, compression)
}
\description{
Write /bids/ group to HDF5 file
}
\keyword{internal}
