% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_h5_write.R
\name{.write_parcellation_group}
\alias{.write_parcellation_group}
\title{Write /parcellation/ group to HDF5 file}
\usage{
.write_parcellation_group(h5, clusters, cluster_ids, compression)
}
\description{
Write /parcellation/ group to HDF5 file
}
\keyword{internal}
