% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids_h5_write.R
\name{.write_spatial_group}
\alias{.write_spatial_group}
\title{Write /spatial/ group to HDF5 file}
\usage{
.write_spatial_group(h5, mask, compression)
}
\description{
Write /spatial/ group to HDF5 file
}
\keyword{internal}
