% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-frame.R
\name{entity-accessors}
\alias{entity-accessors}
\alias{entities}
\alias{entity}
\alias{entity_names}
\title{Access entities from a frame or registry}
\usage{
entities(x, ...)

entity(x, name, ...)

entity_names(x)
}
\arguments{
\item{x}{An \code{fmri_frame}, view, or \code{entity_registry}.}

\item{...}{Additional method arguments.}

\item{name}{One registered entity name.}
}
\value{
\code{entities()} returns the registry; \code{entity()} returns one
\code{entity_frame}; \code{entity_names()} returns registry names.
}
\description{
Access entities from a frame or registry
}
\examples{
subjects <- entity_frame(
  data = tibble::tibble(subject_id = c("sub-1", "sub-2")),
  key = "subject_id"
)
registry <- entity_registry(subject = subjects)
entities(registry)
entity(registry, "subject")
entity_names(registry)
}
