% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri-study.R
\name{event-accessors}
\alias{event-accessors}
\alias{event_data}
\alias{event_key}
\title{Event-table accessors}
\usage{
event_data(x)

event_key(x)
}
\arguments{
\item{x}{An \code{fmri_event_table}.}
}
\value{
Event scalar data or the stable key name.
}
\description{
Event-table accessors
}
\examples{
et <- event_table(
  data.frame(event_id = c("e1", "e2"), onset = c(0, 10)),
  key = "event_id"
)
event_data(et)
event_key(et)
}
