% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_backend_seed.R
\name{extract_array,StudyBackendSeed-method}
\alias{extract_array,StudyBackendSeed-method}
\title{Extract array subset from StudyBackendSeed}
\usage{
\S4method{extract_array}{StudyBackendSeed}(x, index)
}
\description{
This is the key method that enables lazy evaluation. It only loads
the specific subjects and voxels requested.
}
