% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_backend_seed.R
\name{find_subjects_for_rows}
\alias{find_subjects_for_rows}
\title{Find which subjects contain the requested rows}
\usage{
find_subjects_for_rows(rows, boundaries)
}
\description{
Find which subjects contain the requested rows
}
\keyword{internal}
