% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_constructors.R
\name{fmri_latent_dataset}
\alias{fmri_latent_dataset}
\title{Create an fMRI Dataset Object from LatentNeuroVec Files or Objects}
\usage{
fmri_latent_dataset(
  latent_files,
  mask_source = NULL,
  TR,
  run_length,
  event_table = data.frame(),
  base_path = ".",
  censor = NULL,
  preload = FALSE
)
}
\arguments{
\item{latent_files}{Source files or objects}

\item{mask_source}{Ignored}

\item{TR}{The repetition time in seconds}

\item{run_length}{Vector of run lengths}

\item{event_table}{Event table}

\item{base_path}{Base path for files}

\item{censor}{Censor vector}

\item{preload}{Whether to preload data}
}
\value{
A latent_dataset object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Please use \code{latent_dataset()} instead,
which provides a proper interface for latent space data.
}
\examples{
\dontrun{
# Use latent_dataset() instead:
dset <- latent_dataset(
  source = c("run1.lv.h5", "run2.lv.h5", "run3.lv.h5"),
  TR = 2,
  run_length = c(150, 150, 150)
)
}

}
\seealso{
\code{\link{latent_dataset}}
}
