% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_constructors.R
\name{fmri_study_dataset}
\alias{fmri_study_dataset}
\title{Create an fmri_study_dataset}
\usage{
fmri_study_dataset(datasets, subject_ids = NULL)
}
\arguments{
\item{datasets}{A list of \code{fmri_dataset} objects}

\item{subject_ids}{Optional vector of subject identifiers}
}
\value{
An object of class \code{fmri_study_dataset}
}
\description{
High level constructor that combines multiple \code{fmri_dataset} objects
into a single study-level dataset using \code{study_backend}.
}
