% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{fmridataset_error}
\alias{fmridataset_error}
\title{Create a Custom fmridataset Error}
\usage{
fmridataset_error(message, class = character(), ...)
}
\arguments{
\item{message}{Character string describing the error}

\item{class}{Character vector of error classes}

\item{...}{Additional data to include in the error condition}
}
\value{
A condition object
}
\description{
Create a Custom fmridataset Error
}
\examples{
cond <- fmridataset:::fmridataset_error("bad input", class = "fmridataset_error_config")
inherits(cond, "fmridataset_error")
}
\keyword{internal}
