% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FmriSeries.R
\docType{class}
\name{FmriSeries-class}
\alias{FmriSeries-class}
\title{fMRI Time Series Container}
\description{
An S4 class representing lazily accessed fMRI time series data. The
underlying data is stored in a \code{DelayedMatrix} with rows corresponding
to timepoints and columns corresponding to voxels.
}
\section{Slots}{

\describe{
\item{\code{voxel_info}}{A \code{S4Vectors::DataFrame} containing spatial metadata
for each voxel.}

\item{\code{temporal_info}}{A \code{S4Vectors::DataFrame} containing metadata for
each timepoint.}

\item{\code{selection_info}}{A list describing how the data were selected.}

\item{\code{dataset_info}}{A list describing the source dataset and backend.}
}}

