% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_helpers.R
\name{generate_example_events}
\alias{generate_example_events}
\title{Generate Example Event Table}
\usage{
generate_example_events(
  n_runs = 2,
  events_per_run = 4,
  TR = 2,
  run_length = 100
)
}
\arguments{
\item{n_runs}{Number of runs}

\item{events_per_run}{Events per run}

\item{TR}{Repetition time}

\item{run_length}{Length of each run in TRs}
}
\value{
Data frame with event information
}
\description{
Generate Example Event Table
}
\keyword{internal}
