% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_helpers.R
\name{generate_example_mask}
\alias{generate_example_mask}
\title{Create Example Mask}
\usage{
generate_example_mask(n_voxels = 1000, fraction_valid = 0.8)
}
\arguments{
\item{n_voxels}{Total number of voxels}

\item{fraction_valid}{Fraction of voxels to include in mask}
}
\value{
Logical vector mask
}
\description{
Create Example Mask
}
\keyword{internal}
