% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{generics}
\alias{generics}
\title{Generic Functions for fMRI Dataset Operations}
\description{
This file contains all generic function declarations for the refactored
fmridataset package. These establish the interface contracts that are
implemented by dataset-specific methods in other files.
}
