% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_dataset.R
\name{get_component_info}
\alias{get_component_info}
\title{Get Component Information}
\usage{
get_component_info(x, ...)
}
\arguments{
\item{x}{A latent_dataset object}

\item{...}{Additional arguments}
}
\value{
A list containing component metadata
}
\description{
Get metadata about the latent components in the dataset.
}
\seealso{
Other latent_data: 
\code{\link{get_latent_scores}()},
\code{\link{get_spatial_loadings}()},
\code{\link{latent_dataset}()},
\code{\link{reconstruct_voxels}()}
}
\concept{latent_data}
