% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{get_data_matrix}
\alias{get_data_matrix}
\title{Get Data Matrix from fMRI Dataset Objects}
\usage{
get_data_matrix(x, ...)
}
\arguments{
\item{x}{An fMRI dataset object}

\item{...}{Additional arguments passed to methods}
}
\value{
A matrix with timepoints as rows and voxels as columns
}
\description{
Generic function to extract data as a matrix from various fMRI dataset types.
Always returns a matrix with timepoints as rows and voxels as columns.
}
