% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{get_run_duration}
\alias{get_run_duration}
\title{Get run duration from sampling frame}
\usage{
get_run_duration(x, ...)
}
\arguments{
\item{x}{Sampling frame object}

\item{...}{Additional arguments}
}
\description{
Get run duration from sampling frame
}
