% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{get_run_lengths}
\alias{get_run_lengths}
\title{Get run lengths from sampling frame}
\usage{
get_run_lengths(x, ...)
}
\arguments{
\item{x}{Sampling frame object}

\item{...}{Additional arguments}
}
\description{
Get run lengths from sampling frame
}
