% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{get_total_duration}
\alias{get_total_duration}
\title{Get total duration from sampling frame}
\usage{
get_total_duration(x, ...)
}
\arguments{
\item{x}{Sampling frame object}

\item{...}{Additional arguments}
}
\description{
Get total duration from sampling frame
}
