% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{get_TR}
\alias{get_TR}
\title{Get TR from sampling frame}
\usage{
get_TR(x, ...)
}
\arguments{
\item{x}{Sampling frame object}

\item{...}{Additional arguments}
}
\description{
Get TR from sampling frame
}
