% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{global_onsets}
\alias{global_onsets}
\title{Get global onsets from sampling frame}
\usage{
global_onsets(x, ...)
}
\arguments{
\item{x}{Sampling frame object}

\item{...}{Additional arguments}
}
\description{
Get global onsets from sampling frame
}
