% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_backend.R
\name{h5_backend}
\alias{h5_backend}
\title{Create an H5 Backend}
\usage{
h5_backend(
  source,
  mask_source,
  mask_dataset = "data/elements",
  data_dataset = "data/elements",
  preload = FALSE
)
}
\arguments{
\item{source}{Character vector of file paths to H5 files or list of H5NeuroVec objects}

\item{mask_source}{File path to H5 mask file, H5 file containing mask, or in-memory NeuroVol object}

\item{mask_dataset}{Character string specifying the dataset path within H5 file for mask (default: "data/elements")}

\item{data_dataset}{Character string specifying the dataset path within H5 files for data (default: "data/elements")}

\item{preload}{Logical, whether to eagerly load H5NeuroVec objects into memory}
}
\value{
An h5_backend S3 object
}
\description{
Create an H5 Backend
}
\keyword{internal}
