% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hasMethod}
\alias{hasMethod}
\title{Check if a method exists for a given class}
\usage{
hasMethod(generic, class)
}
\arguments{
\item{generic}{Generic function name}

\item{class}{Class name}
}
\description{
Internal helper to check S3 method existence
}
\keyword{internal}
