% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FmriSeries.R
\name{is.fmri_series}
\alias{is.fmri_series}
\title{Check if object is an fmri_series}
\usage{
is.fmri_series(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
Logical TRUE if x is an fmri_series object
}
\description{
Check if object is an fmri_series
}
