% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_frame_adapters.R
\name{is.sampling_frame}
\alias{is.sampling_frame}
\title{Test if Object is a Sampling Frame}
\usage{
is.sampling_frame(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
TRUE if x is a sampling_frame object, FALSE otherwise
}
\description{
This function tests whether an object is of class 'sampling_frame'.
}
