% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_backend_seed.R
\name{is_sparse,StudyBackendSeed-method}
\alias{is_sparse,StudyBackendSeed-method}
\title{Check if object is sparse (always FALSE for fMRI data)}
\usage{
\S4method{is_sparse}{StudyBackendSeed}(x)
}
\description{
Check if object is sparse (always FALSE for fMRI data)
}
