% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_standards.R
\name{mask_to_logical}
\alias{mask_to_logical}
\title{Convert Mask to Logical Vector}
\usage{
mask_to_logical(mask)
}
\arguments{
\item{mask}{A mask in any supported format (numeric, logical, array, NeuroVol)}
}
\value{
A logical vector
}
\description{
Converts any mask representation to a logical vector.
}
\keyword{internal}
