% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_standards.R
\name{mask_to_volume}
\alias{mask_to_volume}
\title{Convert Logical Vector to Volume}
\usage{
mask_to_volume(mask_vec, dims)
}
\arguments{
\item{mask_vec}{A logical vector}

\item{dims}{Spatial dimensions (length 3)}
}
\value{
A 3D logical array
}
\description{
Converts a logical vector mask to a 3D array.
}
\keyword{internal}
