% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{n_runs}
\alias{n_runs}
\title{Get number of runs from sampling frame}
\usage{
n_runs(x, ...)
}
\arguments{
\item{x}{Sampling frame object}

\item{...}{Additional arguments}
}
\description{
Get number of runs from sampling frame
}
