% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_group.R
\name{n_subjects}
\alias{n_subjects}
\title{Number of subjects in a group}
\usage{
n_subjects(gd)
}
\arguments{
\item{gd}{An \code{fmri_group}.}
}
\value{
Integer number of subjects.
}
\description{
Number of subjects in a group
}
