% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{n_timepoints}
\alias{n_timepoints}
\title{Get number of timepoints from sampling frame}
\usage{
n_timepoints(x, ...)
}
\arguments{
\item{x}{Sampling frame object}

\item{...}{Additional arguments}
}
\description{
Get number of timepoints from sampling frame
}
