% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FmriSeries.R
\name{ncol.fmri_series}
\alias{ncol.fmri_series}
\title{Number of columns in fmri_series}
\usage{
\method{ncol}{fmri_series}(x)
}
\arguments{
\item{x}{An fmri_series object}
}
\value{
Number of voxels
}
\description{
Number of columns in fmri_series
}
