% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FmriSeries.R
\name{nrow.fmri_series}
\alias{nrow.fmri_series}
\title{Number of rows in fmri_series}
\usage{
\method{nrow}{fmri_series}(x)
}
\arguments{
\item{x}{An fmri_series object}
}
\value{
Number of timepoints
}
\description{
Number of rows in fmri_series
}
