% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/bids_h5_dataset.R
\name{parcellation_info}
\alias{parcellation_info}
\alias{parcellation_info.bids_h5_study_dataset}
\title{Get Parcellation Information from a BIDS H5 Dataset}
\usage{
parcellation_info(x, ...)

\method{parcellation_info}{bids_h5_study_dataset}(x, ...)
}
\arguments{
\item{x}{A BIDS H5 study dataset object}

\item{...}{Additional arguments passed to methods}
}
\value{
A list with elements: cluster_ids, cluster_map, labels, n_parcels
}
\description{
Generic function to retrieve parcellation metadata from a BIDS HDF5
dataset stored in parcellated mode.
}
