% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.fmri_dataset}
\alias{print.fmri_dataset}
\alias{summary.fmri_dataset}
\alias{print.latent_dataset}
\alias{print.chunkiter}
\alias{print.data_chunk}
\alias{print.matrix_dataset}
\title{Pretty Print a Chunk Iterator}
\usage{
\method{print}{fmri_dataset}(x, full = FALSE, ...)

\method{summary}{fmri_dataset}(object, ...)

\method{print}{latent_dataset}(x, ...)

\method{print}{chunkiter}(x, ...)

\method{print}{data_chunk}(x, ...)

\method{print}{matrix_dataset}(x, ...)
}
\arguments{
\item{x}{A data_chunk object.}

\item{...}{Additional arguments (ignored).}
}
\description{
This function prints a summary of a chunk iterator using colored output.

This function prints a summary of a data chunk using crayon for colored output.
}
