% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_dcf_config}
\alias{read_dcf_config}
\title{Read DCF-style configuration (backwards compatibility)}
\usage{
read_dcf_config(file_name)
}
\description{
Read DCF-style configuration (backwards compatibility)
}
\keyword{internal}
