% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_backend_seed_s3.R
\name{register_study_backend_seed_methods}
\alias{register_study_backend_seed_methods}
\title{Register S3 Methods for DelayedArray Compatibility}
\usage{
register_study_backend_seed_methods()
}
\description{
This function should be called in .onLoad to register the S3 methods
with DelayedArray's S4 generic functions.
}
\keyword{internal}
