% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{samples}
\alias{samples}
\title{Get samples from sampling frame}
\usage{
samples(x, ...)
}
\arguments{
\item{x}{Sampling frame object}

\item{...}{Additional arguments}
}
\description{
Get samples from sampling frame
}
