% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame-view.R
\name{select_features}
\alias{select_features}
\title{Select frame features using feature metadata}
\usage{
select_features(x, predicate)
}
\arguments{
\item{x}{An \code{fmri_frame} or view.}

\item{predicate}{A metadata expression returning one logical value per
feature.}
}
\value{
An \code{fmri_view}.
}
\description{
Select frame features using feature metadata
}
\examples{
sp <- volume_space(dim = c(2, 2, 2), affine = diag(4))
frame <- fmri_frame(
  assays = list(bold = matrix(rnorm(4 * n_features(sp)), nrow = 4)),
  observations = data.frame(.obs_id = sprintf("vol-\%d", 1:4)),
  space = sp
)
select_features(frame, i == 1)
}
