% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri-frame.R
\name{space}
\alias{space}
\title{Feature-space accessor}
\usage{
space(x, ...)
}
\arguments{
\item{x}{An object with spatial identity.}

\item{...}{Additional arguments.}
}
\value{
A \code{FeatureSpace}.
}
\description{
Feature-space accessor
}
\examples{
sp <- volume_space(dim = c(2, 2, 2), affine = diag(4))
frame <- fmri_frame(
  assays = list(bold = matrix(rnorm(4 * n_features(sp)), nrow = 4)),
  observations = data.frame(.obs_id = sprintf("vol-\%d", 1:4)),
  space = sp
)
space(frame)
}
