% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri-frame.R
\name{spatial_map}
\alias{spatial_map}
\title{Recover a spatial map for one observation}
\usage{
spatial_map(x, observation, assay = active_assay(x))
}
\arguments{
\item{x}{An \code{fmri_frame} or view.}

\item{observation}{Observation ID or one integer position.}

\item{assay}{Assay name.}
}
\value{
A reconstructed spatial object.
}
\description{
Recover a spatial map for one observation
}
\examples{
sp <- volume_space(dim = c(2, 2, 2), affine = diag(4))
frame <- fmri_frame(
  assays = list(bold = matrix(rnorm(4 * n_features(sp)), nrow = 4)),
  observations = data.frame(.obs_id = sprintf("vol-\%d", 1:4)),
  space = sp
)
spatial_map(frame, observation = 1L)
}
