% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_selector.R
\name{sphere_selector}
\alias{sphere_selector}
\title{Spherical ROI Series Selector}
\usage{
sphere_selector(center, radius)
}
\arguments{
\item{center}{Numeric vector of length 3 (x, y, z) specifying sphere center}

\item{radius}{Numeric radius in voxel units}
}
\value{
An object of class \code{sphere_selector}
}
\description{
Select voxels within a spherical region.
}
\examples{
# Select 10-voxel radius sphere around voxel (30, 30, 20)
sel <- sphere_selector(center = c(30, 30, 20), radius = 10)
}
