% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri-study.R
\name{study-accessors}
\alias{study-accessors}
\alias{study_frames}
\alias{study_frame}
\alias{study_ids}
\title{Study representation accessors}
\usage{
study_frames(x, contextual = TRUE)

study_frame(x, name, contextual = TRUE)

study_ids(x)
}
\arguments{
\item{x}{An \code{fmri_study}.}

\item{contextual}{Replace frame-local entity stubs with shared study entities.}

\item{name}{Stable representation name.}
}
\value{
Named representations, one representation, or representation IDs.
}
\description{
Study representation accessors
}
\examples{
voxels <- volume_space(dim = c(2, 2, 1), affine = diag(4), template = "toy")
frame <- fmri_frame(
  assays = list(bold = matrix(rnorm(3 * n_features(voxels)), nrow = 3)),
  observations = data.frame(.obs_id = paste0("vol-", 1:3)),
  space = voxels
)
study <- fmri_study(list(main = frame))
study_ids(study)
study_frame(study, "main")
}
