% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri-study.R
\name{study-registries}
\alias{study-registries}
\alias{study_links}
\alias{study_link}
\alias{study_tables}
\alias{study_table}
\alias{events}
\title{Study link and table accessors}
\usage{
study_links(x)

study_link(x, name)

study_tables(x)

study_table(x, name)

events(x, name = "events")
}
\arguments{
\item{x}{An \code{fmri_study}.}

\item{name}{Stable link or table name.}
}
\value{
A registry or one descriptor/table.
}
\description{
Study link and table accessors
}
\examples{
voxels <- volume_space(dim = c(2, 2, 1), affine = diag(4), template = "toy")
frame <- fmri_frame(
  assays = list(bold = matrix(rnorm(3 * n_features(voxels)), nrow = 3)),
  observations = data.frame(.obs_id = paste0("vol-", 1:3)),
  space = voxels
)
study <- fmri_study(list(main = frame))
study_links(study)
study_tables(study)
}
