% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_backend_seed_s3.R
\name{study_backend_seed}
\alias{study_backend_seed}
\title{Create a Study Backend Seed}
\usage{
study_backend_seed(backends, subject_ids)
}
\arguments{
\item{backends}{List of subject backend objects}

\item{subject_ids}{Character vector of subject identifiers}
}
\value{
A StudyBackendSeed S4 object
}
\description{
Create a Study Backend Seed
}
\keyword{internal}
