% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_backend_seed_s3.R
\name{study_seed_chunk_grid}
\alias{study_seed_chunk_grid}
\title{Get Chunk Grid for Study Backend Seed}
\usage{
study_seed_chunk_grid(x, chunk_dim = NULL)
}
\arguments{
\item{x}{A study_backend_seed object}

\item{chunk_dim}{Optional chunk dimensions}
}
\value{
A RegularArrayGrid object
}
\description{
Get Chunk Grid for Study Backend Seed
}
\keyword{internal}
