% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_backend_seed_s3.R
\name{study_seed_is_sparse}
\alias{study_seed_is_sparse}
\title{Check if Study Backend Seed is Sparse}
\usage{
study_seed_is_sparse(x)
}
\arguments{
\item{x}{A study_backend_seed object}
}
\value{
Logical indicating if the data is sparse
}
\description{
Check if Study Backend Seed is Sparse
}
\keyword{internal}
