% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_backend_seed.R
\name{StudyBackendSeed}
\alias{StudyBackendSeed}
\title{Constructor for StudyBackendSeed}
\usage{
StudyBackendSeed(backends, subject_ids)
}
\description{
Constructor for StudyBackendSeed
}
\keyword{internal}
