% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/dataset_methods.R
\name{subject_ids}
\alias{subject_ids}
\alias{subject_ids.fmri_study_dataset}
\title{Get Subject IDs from Multi-Subject Dataset}
\usage{
subject_ids(x, ...)

\method{subject_ids}{fmri_study_dataset}(x, ...)
}
\arguments{
\item{x}{A multi-subject dataset object (e.g., fmri_study_dataset)}

\item{...}{Additional arguments passed to methods}
}
\value{
Character vector of subject identifiers
}
\description{
Generic function to extract subject identifiers from multi-subject
fMRI dataset objects.
}
\details{
Multi-subject datasets contain data from multiple participants. This
function extracts the subject identifiers associated with each dataset.
The order of subject IDs corresponds to the order of datasets.
}
