% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri_group.R
\name{validate_fmri_group}
\alias{validate_fmri_group}
\title{Validate an fmri_group object}
\usage{
validate_fmri_group(x)
}
\arguments{
\item{x}{Object to validate.}
}
\value{
The validated \code{fmri_group} (invisibly).
}
\description{
Validate an fmri_group object
}
