% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri-study.R
\name{validate_fmri_study}
\alias{validate_fmri_study}
\title{Validate a study}
\usage{
validate_fmri_study(x)
}
\arguments{
\item{x}{An \code{fmri_study}.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate a study
}
\examples{
voxels <- volume_space(dim = c(2, 2, 1), affine = diag(4), template = "toy")
frame <- fmri_frame(
  assays = list(bold = matrix(rnorm(3 * n_features(voxels)), nrow = 3)),
  observations = data.frame(.obs_id = paste0("vol-", 1:3)),
  space = voxels
)
study <- fmri_study(list(main = frame))
validate_fmri_study(study)
}
