% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_selector.R
\name{voxel_selector}
\alias{voxel_selector}
\title{Voxel Coordinate Series Selector}
\usage{
voxel_selector(coords)
}
\arguments{
\item{coords}{Matrix with 3 columns (x, y, z) or vector of length 3}
}
\value{
An object of class \code{voxel_selector}
}
\description{
Select voxels by their 3D coordinates in the image space.
}
\examples{
# Select single voxel
sel <- voxel_selector(c(10, 20, 15))

# Select multiple voxels
coords <- cbind(x = c(10, 20), y = c(20, 30), z = c(15, 15))
sel <- voxel_selector(coords)
}
