% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri-frame.R
\name{aligned_assay_set}
\alias{aligned_assay_set}
\title{Construct a strictly aligned assay set}
\usage{
aligned_assay_set(assays, observations, features)
}
\arguments{
\item{assays}{Named matrices or \code{ArraySource} descriptors.}

\item{observations}{Observation axis.}

\item{features}{Feature axis.}
}
\value{
An \code{aligned_assay_set}.
}
\description{
Construct a strictly aligned assay set
}
