% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-source.R
\name{array-source}
\alias{array-source}
\alias{as_array_source}
\alias{source_shape}
\alias{source_dtype}
\alias{source_chunks}
\alias{source_capabilities}
\alias{source_fingerprint}
\alias{source_open}
\alias{source_read}
\alias{source_read_native}
\alias{source_close}
\title{Serializable numerical array sources}
\usage{
as_array_source(x, ...)

source_shape(x, ...)

source_dtype(x, ...)

source_chunks(x, ...)

source_capabilities(x, ...)

source_fingerprint(x, ...)

source_open(x, ...)

source_read(x, observations = NULL, features = NULL, ...)

source_read_native(x, observations = NULL, ...)

source_close(x, ...)
}
\arguments{
\item{x}{An array source or object coercible to one.}

\item{...}{Additional method arguments.}

\item{observations}{Optional observation positions.}

\item{features}{Optional feature positions.}
}
\description{
Serializable numerical array sources
}
