% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FmriSeries.R
\name{as.matrix.FmriSeries}
\alias{as.matrix.FmriSeries}
\title{Materialise an FmriSeries as a standard matrix}
\usage{
\method{as.matrix}{FmriSeries}(x, ...)
}
\arguments{
\item{x}{An \code{FmriSeries} object}

\item{...}{Additional arguments (ignored)}
}
\value{
A matrix
}
\description{
This method realises the underlying DelayedMatrix and
returns an ordinary matrix with timepoints in rows and
voxels in columns.
}
