% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{as.matrix_dataset}
\alias{as.matrix_dataset}
\title{Convert to Matrix Dataset}
\usage{
as.matrix_dataset(x, ...)
}
\arguments{
\item{x}{An fMRI dataset object}

\item{...}{Additional arguments passed to methods}
}
\value{
A matrix_dataset object
}
\description{
Generic function to convert various fMRI dataset types to matrix_dataset objects.
Provides a unified interface for getting matrix-based representations.
}
