% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FmriSeries.R
\name{as_tibble.FmriSeries}
\alias{as_tibble.FmriSeries}
\title{Convert an FmriSeries to a tidy tibble}
\usage{
\method{as_tibble}{FmriSeries}(x, ...)
}
\arguments{
\item{x}{An \code{FmriSeries} object}

\item{...}{Additional arguments (ignored)}
}
\value{
A tibble with signal and metadata columns
}
\description{
The returned tibble contains one row per voxel/timepoint
combination with metadata columns from \code{temporal_info}
and \code{voxel_info} and a \code{signal} column with the data
values.
}
