% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_backend.R, R/latent_backend.R,
%   R/matrix_backend.R, R/nifti_backend.R, R/storage_backend.R,
%   R/study_backend.R, R/zarr_backend.R
\name{backend_get_dims.h5_backend}
\alias{backend_get_dims.h5_backend}
\alias{backend_get_dims.latent_backend}
\alias{backend_get_dims.matrix_backend}
\alias{backend_get_dims.nifti_backend}
\alias{backend_get_dims}
\alias{backend_get_dims.study_backend}
\alias{backend_get_dims.zarr_backend}
\title{Get Dimensions from Backend}
\usage{
\method{backend_get_dims}{h5_backend}(backend)

\method{backend_get_dims}{latent_backend}(backend)

\method{backend_get_dims}{matrix_backend}(backend)

\method{backend_get_dims}{nifti_backend}(backend)

backend_get_dims(backend)

\method{backend_get_dims}{study_backend}(backend)

\method{backend_get_dims}{zarr_backend}(backend)
}
\arguments{
\item{backend}{A storage backend object}
}
\value{
A named list with elements:
\itemize{
\item spatial: numeric vector of length 3 (x, y, z dimensions)
\item time: integer, number of timepoints
}
}
\description{
Returns the dimensions of the data stored in the backend.
}
\keyword{internal}
