% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri-frame.R
\name{bind_observations}
\alias{bind_observations}
\title{Bind frames along observations}
\usage{
bind_observations(...)
}
\arguments{
\item{...}{Frames with identical feature IDs, spaces, and assay names.}
}
\value{
A lazily row-bound \code{fmri_frame}.
}
\description{
Bind frames along observations
}
