% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri-frame.R
\name{block_apply}
\alias{block_apply}
\title{Apply a function to bounded feature blocks}
\usage{
block_apply(x, FUN, block_size = 4096L, assay = active_assay(x), ...)
}
\arguments{
\item{x}{An \code{fmri_frame} or view.}

\item{FUN}{Function receiving an observation-by-feature matrix and feature IDs.}

\item{block_size}{Number of features per block.}

\item{assay}{Assay name.}

\item{...}{Additional arguments passed to \code{FUN}.}
}
\value{
A list of block results.
}
\description{
Apply a function to bounded feature blocks
}
