% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chunks.R
\name{collect_chunks}
\alias{collect_chunks}
\title{Collect all chunks from a chunk iterator}
\usage{
collect_chunks(chunk_iter)
}
\arguments{
\item{chunk_iter}{A chunk iterator object created by chunk_iter()}
}
\value{
A list containing all chunks from the iterator
}
\description{
This function collects all chunks from a chunk iterator into a list.
}
