% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_methods.R
\name{dataset_methods}
\alias{dataset_methods}
\title{Dataset Methods for fmridataset}
\description{
This file implements methods for dataset objects that delegate to
their internal sampling_frame objects for temporal information.
}
\keyword{internal}
