% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-frame.R
\name{entity-frame-accessors}
\alias{entity-frame-accessors}
\alias{entity_key}
\alias{entity_ids}
\alias{entity_data}
\alias{entity_blocks}
\title{Entity-frame accessors}
\usage{
entity_key(x)

entity_ids(x)

entity_data(x)

entity_blocks(x)
}
\arguments{
\item{x}{An \code{entity_frame}.}
}
\value{
The stable key name, entity IDs, scalar data, or aligned blocks.
}
\description{
Entity-frame accessors
}
