% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-space.R
\name{feature-space}
\alias{feature-space}
\alias{n_features}
\alias{feature_ids}
\alias{native_shape}
\alias{feature_data}
\alias{space_digest}
\alias{restrict_space}
\alias{vectorize_space}
\alias{reconstruct_space}
\alias{adjacency}
\alias{compatible_space}
\alias{assert_compatible_space}
\alias{reconstruct_space.surface_space}
\title{Feature-space contract}
\usage{
n_features(x, ...)

feature_ids(x, ...)

native_shape(x, ...)

feature_data(x, ...)

space_digest(x, ...)

restrict_space(x, index, ...)

vectorize_space(x, spatial_object, ...)

reconstruct_space(x, vector, ...)

adjacency(x, ...)

compatible_space(x, y, ...)

assert_compatible_space(x, y, ...)

\method{reconstruct_space}{surface_space}(x, vector, format = c("surface_map", "neurosurf"), ...)
}
\arguments{
\item{x}{A feature-space object.}

\item{...}{Additional arguments for methods.}

\item{index}{Feature positions used to restrict a space.}

\item{spatial_object}{A native spatial object to vectorize.}

\item{vector}{A feature vector to reconstruct.}

\item{y}{A second feature-space object.}

\item{format}{Surface reconstruction format. The backend-neutral default is
\code{"surface_map"}; \code{"neurosurf"} returns a \code{neurosurf::NeuroSurface} when
embedded unilateral geometry is available.}
}
\description{
Feature-space contract
}
