% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_backend_seed_s3.R
\name{find_subjects_for_rows}
\alias{find_subjects_for_rows}
\title{Find Which Subjects Contain Given Rows}
\usage{
find_subjects_for_rows(rows, boundaries)
}
\description{
Find Which Subjects Contain Given Rows
}
\keyword{internal}
