% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri-frame.R, R/frame-view.R
\name{frame-accessors}
\alias{frame-accessors}
\alias{assays}
\alias{assay}
\alias{active_assay}
\alias{observation_axis}
\alias{observations}
\alias{features}
\alias{observation_ids}
\alias{obs_blocks}
\alias{feature_blocks}
\alias{dim.fmri_frame}
\alias{nrow.fmri_frame}
\alias{ncol.fmri_frame}
\alias{dim.fmri_view}
\alias{nrow.fmri_view}
\alias{ncol.fmri_view}
\title{Frame accessors}
\usage{
assays(x, ...)

assay(x, name = active_assay(x), ...)

active_assay(x, ...)

observation_axis(x, ...)

observations(x, ...)

features(x, ...)

observation_ids(x, ...)

obs_blocks(x, ...)

feature_blocks(x, ...)

\method{dim}{fmri_frame}(x)

nrow.fmri_frame(x)

ncol.fmri_frame(x)

\method{dim}{fmri_view}(x)

nrow.fmri_view(x)

ncol.fmri_view(x)
}
\arguments{
\item{x}{An \code{fmri_frame} or \code{fmri_view}.}

\item{...}{Additional method arguments.}

\item{name}{Assay name.}
}
\description{
Frame accessors
}
