% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{get_data}
\alias{get_data}
\title{Get Data from fMRI Dataset Objects}
\usage{
get_data(x, ...)
}
\arguments{
\item{x}{An fMRI dataset object}

\item{...}{Additional arguments passed to methods}
}
\value{
Dataset-specific data object
}
\description{
Generic function to extract data from various fMRI dataset types.
Returns the underlying data in its native format (NeuroVec, matrix, etc.).
}
