% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{get_mask}
\alias{get_mask}
\title{Get Mask from fMRI Dataset Objects}
\usage{
get_mask(x, ...)
}
\arguments{
\item{x}{An fMRI dataset object}

\item{...}{Additional arguments passed to methods}
}
\value{
Mask object (NeuroVol, vector, etc.)
}
\description{
Generic function to extract masks from various fMRI dataset types.
Returns the mask in its appropriate format for the dataset type.
}
