% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-space.R
\name{index_space}
\alias{index_space}
\title{Construct a generic indexed feature space}
\usage{
index_space(n, ids = NULL, namespace = NULL, data = NULL)
}
\arguments{
\item{n}{Number of features.}

\item{ids}{Optional stable feature IDs.}

\item{namespace}{Namespace used for generated IDs.}

\item{data}{Optional feature metadata.}
}
\value{
An \code{index_space}.
}
\description{
Construct a generic indexed feature space
}
