% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-source.R
\name{memory_source}
\alias{memory_source}
\title{Construct an in-memory array source}
\usage{
memory_source(data, dtype = NULL, chunks = NULL)
}
\arguments{
\item{data}{A two-dimensional matrix or array.}

\item{dtype}{Logical storage dtype. Numeric R matrices default to
\code{"float64"}.}

\item{chunks}{Optional logical chunk shape.}
}
\value{
A serializable \code{memory_source}.
}
\description{
Construct an in-memory array source
}
