% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/dataset_methods.R,
%   R/sampling_frame_adapters.R
\name{n_timepoints}
\alias{n_timepoints}
\alias{n_timepoints.matrix_dataset}
\alias{n_timepoints.fmri_dataset}
\alias{n_timepoints.fmri_mem_dataset}
\alias{n_timepoints.fmri_file_dataset}
\alias{n_timepoints.fmri_study_dataset}
\alias{n_timepoints.sampling_frame}
\title{Get Number of Timepoints from Sampling Frame}
\usage{
n_timepoints(x, ...)

\method{n_timepoints}{matrix_dataset}(x, ...)

\method{n_timepoints}{fmri_dataset}(x, ...)

\method{n_timepoints}{fmri_mem_dataset}(x, ...)

\method{n_timepoints}{fmri_file_dataset}(x, ...)

\method{n_timepoints}{fmri_study_dataset}(x, ...)

\method{n_timepoints}{sampling_frame}(x, ...)
}
\arguments{
\item{x}{An object containing temporal structure (e.g., sampling_frame, fmri_dataset)}

\item{...}{Additional arguments passed to methods}
}
\value{
Integer representing the total number of timepoints
}
\description{
Extracts the total number of timepoints (volumes) across all runs
from objects containing temporal structure information.
}
\examples{
\donttest{
# Create a sampling frame with 3 runs
sf <- fmrihrf::sampling_frame(blocklens = c(100, 120, 110), TR = 2)
n_timepoints(sf) # Returns: 330 (sum of run lengths)
}

}
\seealso{
\code{\link{n_runs}} for number of runs,
\code{\link{get_run_lengths}} for individual run lengths
}
