% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-validity.R
\name{observation_validity}
\alias{observation_validity}
\title{Resolve validity onto frame observations}
\usage{
observation_validity(x, name = NULL)
}
\arguments{
\item{x}{An \code{fmri_frame} or view.}

\item{name}{Validity relation name.}
}
\value{
Observation-by-feature logical validity matrix.
}
\description{
Resolve validity onto frame observations
}
