% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_fmri_config}
\alias{read_fmri_config}
\title{read a basic fMRI configuration file}
\usage{
read_fmri_config(file_name, base_path = NULL)
}
\arguments{
\item{file_name}{name of configuration file}

\item{base_path}{Base directory for relative file names. Absolute paths are used as-is.}
}
\value{
a \code{fmri_config} instance
}
\description{
read a basic fMRI configuration file
}
