% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relation-registry.R
\name{relation_registry}
\alias{relation_registry}
\title{Construct a relation registry}
\usage{
relation_registry(relations = list(), ...)
}
\arguments{
\item{relations}{A named list of \code{key_relation} or \code{sparse_relation}
descriptors.}

\item{...}{Alternatively, named relation descriptors.}
}
\value{
A named \code{relation_registry}.
}
\description{
Construct a relation registry
}
