% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{resolve_indices}
\alias{resolve_indices}
\title{Resolve indices from series selector}
\usage{
resolve_indices(selector, dataset, ...)
}
\arguments{
\item{selector}{Series selector object}

\item{dataset}{Dataset object for context}

\item{...}{Additional arguments}
}
\description{
Resolve indices from series selector
}
