% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-source.R
\name{row_bound_source}
\alias{row_bound_source}
\title{Bind compatible sources along observations}
\usage{
row_bound_source(sources)
}
\arguments{
\item{sources}{A non-empty list of two-dimensional array sources.}
}
\value{
A serializable row-bound source.
}
\description{
Bind compatible sources along observations
}
