% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/dataset_methods.R,
%   R/sampling_frame_adapters.R
\name{samples}
\alias{samples}
\alias{samples.fmri_dataset}
\alias{samples.sampling_frame}
\title{Get Sample Indices from Sampling Frame}
\usage{
samples(x, ...)

\method{samples}{fmri_dataset}(x, ...)

\method{samples}{sampling_frame}(x, ...)
}
\arguments{
\item{x}{An object containing temporal structure (e.g., sampling_frame, fmri_dataset)}

\item{...}{Additional arguments passed to methods}
}
\value{
Integer vector from 1 to the total number of timepoints
}
\description{
Generates a vector of timepoint indices, typically used for
time series analysis or indexing operations.
}
\examples{
\donttest{
# Create a sampling frame
sf <- fmrihrf::sampling_frame(blocklens = c(100, 120), TR = 2)
s <- samples(sf)
length(s) # 220
range(s) # c(1, 220)
}

}
\seealso{
\code{\link{n_timepoints}} for total number of samples,
\code{\link{blockids}} for run membership
}
