% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-source.R
\name{source_view}
\alias{source_view}
\title{Construct a lazy view over an array source}
\usage{
source_view(source, observations = NULL, features = NULL)
}
\arguments{
\item{source}{An \code{array_source}.}

\item{observations}{Stored observation selector.}

\item{features}{Stored feature selector.}
}
\value{
A serializable source view.
}
\description{
Construct a lazy view over an array source
}
