% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri-frame.R
\name{space}
\alias{space}
\title{Feature-space accessor}
\usage{
space(x, ...)
}
\arguments{
\item{x}{An object with spatial identity.}

\item{...}{Additional arguments.}
}
\value{
A \code{FeatureSpace}.
}
\description{
Feature-space accessor
}
