% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{stop_fmridataset}
\alias{stop_fmridataset}
\title{Stop with a Custom Error}
\usage{
stop_fmridataset(error_fn, message = NULL, ...)
}
\arguments{
\item{error_fn}{Error constructor function}

\item{message}{Error message (optional if provided as first ... argument)}

\item{...}{Arguments passed to the error constructor}
}
\description{
Stop with a Custom Error
}
\keyword{internal}
