% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri-study.R
\name{study-registries}
\alias{study-registries}
\alias{study_links}
\alias{study_link}
\alias{study_tables}
\alias{study_table}
\alias{events}
\title{Study link and table accessors}
\usage{
study_links(x)

study_link(x, name)

study_tables(x)

study_table(x, name)

events(x, name = "events")
}
\arguments{
\item{x}{An \code{fmri_study} or filtered view.}

\item{name}{Stable link or table name.}
}
\value{
A registry or one descriptor/table.
}
\description{
Study link and table accessors
}
