% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri-study.R
\name{validate_event_table}
\alias{validate_event_table}
\title{Validate a keyed event table}
\usage{
validate_event_table(x)
}
\arguments{
\item{x}{An \code{fmri_event_table}.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate a keyed event table
}
