% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri-study.R
\name{validate_fmri_study}
\alias{validate_fmri_study}
\title{Validate a study}
\usage{
validate_fmri_study(x)
}
\arguments{
\item{x}{An \code{fmri_study}.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate a study
}
