% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmri-study.R
\name{validate_fmri_study}
\alias{validate_fmri_study}
\title{Validate a study or filtered study view}
\usage{
validate_fmri_study(x)
}
\arguments{
\item{x}{An \code{fmri_study} or \code{fmri_study_view}.}
}
\value{
\code{x}, invisibly.
}
\description{
Validate a study or filtered study view
}
